+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Thanks to the following contributors who worked on this release:
 ### Fixed
 - Fixed an error with importing `.gpx` or `.gp` files with artificial harmonics for which the harmonic type was set to `Pinch`
 - Fixed a MIDI playback bug where let rings could ring forever in certain situations
+- Added a b6 alteration for chord names (#370)
 
 ## [Alpha 17] - 2022-01-28